Tugas Akhir/requirements.txt
flask
torch
torchvision
faiss-cpu
pandas
Pillow
matplotlib
kagglehub
git+https://github.com/openai/CLIP.git
